Codegen v2: contract directory, schema-first ClientState, auto-fixtures
- types/api/ is now the client-contract directory: every exported Zod
  schema in it transpiles to Swift; new files must be registered
  (enforced at codegen time). websocket/user-sessions/attachments moved in.
- ClientState is schema-first: the server type derives from
  ClientStateSchema (createdAt is an ISO string; AI SDK parts stay typed
  via a wireOpaque pass-through), deleting the hand-written duplicate and
  its drift check. Setup-run types converted to Zod with identical shapes.
- Fixtures are auto-synthesized from the JSON Schemas (full/minimal per
  struct, one per union variant; parse-gated) into AutoFixtures.json —
  adding a type to an existing contract file needs no other edits.
  Hand-written fixtures trimmed to realism cases. 233 Swift tests.
- Review fixes: const/default propagation through .nullable(), loud
  error on non-discriminated unions, nested-name collision guards,
  acronym-aware case naming, canonicalized JSON fidelity comparison,
  integer fields now z.number().int() → Int in Swift.

+++ packages/shared/src/types/api/auth.ts
@@ -72,7 +72,7 @@ export const OpenAIDeviceStartResponse = z.object({
   attemptId: z.string(),
   verificationUrl: z.string(),
   userCode: z.string(),
-  intervalSeconds: z.number(),
+  intervalSeconds: z.number().int(),
   expiresAt: z.string(),
 });
 export type OpenAIDeviceStartResponse = z.infer<typeof OpenAIDeviceStartResponse>;
